requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
diff-match-patch==20241021
Django==5.1.7
django-cors-headers==4.7.0
django-import-export==4.3.7
django-jazzmin==3.0.1
djangorestframework==3.15.2
idna==3.10
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
tablib==3.8.0
urllib3==2.3.0
